play a bit more with debug and REPL
- to debug the main application, we need to use the green arrow in
  the margin otherwise the execution is not correctly managed by
intellij. And then we can run the same thing in debug (but direct
shortcuts ctrl + shift + R or ctrl + shift + D) does not work.
- for the REPL, we need to launch it in debug mode, then run "switch
  NS ..." (it will execute `(in-ns 'NAMESPACE)` in the REPL), and
then we can add breakpoint, and just run the function in the REPL,
for example `(add 2 4)` and it will correctly stop at the
breakpoint.

A problem though, this is working only for the namespace
sandbox.core, I don't know why this is not working with the other
namespaces :/