fix compojure handler, don't use map
So let's use basic string response for now.

The server can be run using:
```
lein ring server
```

Then we can curl the endpoint:
```
curl localhost:3000/hello
```

And bonus, the server is doing hot-reload! (it might be good to find
a way to activate some logs, to see that the server has taken the
new sources into account).

🎉